+++ tests/variable_scoping.rs
@@ -7,17 +7,17 @@ fn variable_in_expand() {
     let s = common::Setup::new("csv", r#"
 bar := test + 1
 ---
-Foo,Bar,Baz
-![[e=3]][[var=test]],=test*5,=bar
+Foo,Bar,Baz,=SUM(test)
+![[e=3]][[var=test]],=test*5,=bar,
 "#);
     let template = Template::compile(&s.runtime).unwrap();
     let target = s.runtime.target().unwrap();
     target.write(&template).unwrap();
     
     assert_eq!(s.read_output(), 
-"Foo,Bar,Baz
-,=(A2 * 5),=(A2 + 1)
-,=(A3 * 5),=(A3 + 1)
-,=(A4 * 5),=(A4 + 1)
+"Foo,Bar,Baz,=SUM(A2:A4)
+,=(A2 * 5),=(A2 + 1),
+,=(A3 * 5),=(A3 + 1),
+,=(A4 * 5),=(A4 + 1),
 ");
 }